+++ editors/neovim/queries/mcc/highlights.scm
@@ -50,6 +50,7 @@
   "let"
   "defer"
   "in"
+  "unreachable"
 ] @keyword
 
 [